Allow people and shift lines to differ by two.
Due to trimming newlines, the raw people data may have
less lines than the shift data. The latter include
tabs for the (missing) shifts, so only the last newline
is removed by trim().

The people can be removed up to the last Dummy name, if
there are no attributes set. This happens in older plans.